Use relative paths for manifest tests
